support over-registration
Allow unlimited purchases, putting the ones above the threshold in a waiting queue.

When a user registers more tickets than there are available, the entire order is pushed
on a waiting queue. Inform the user of this, starting from 10 or less available tickets.
In the backend, allow de-queueing. Needs mails.