move from reservations to vouchers
Instead of reservations, use vouchers, which can be generated from the renamed
admin page (/admin/vouchers), and adapt the form so that we show the "you can't
purchase right now" depending on what was entered as a voucher code instead of
the email address.

The voucher named "default" is added upon database creation and is used as
the fallback for any uknown voucher (including when no voucher is specified),
which means that this "default" voucher effectively controls the date from
which voucherless orders can go through.